Refactor crossmatch algorithms
